+++ code/qcommon/msg.c
@@ -781,7 +781,7 @@ typedef struct {
 } netField_t;
 
 // using the stringizing operator to save typing...
-#define	NETF(x) #x,(int)&((entityState_t*)0)->x
+#define	NETF(x) #x,(int)(intptr_t)&((entityState_t*)0)->x
 
 netField_t	entityStateFields[] = 
 {
@@ -1096,7 +1096,7 @@ plyer_state_t communication
 */
 
 // using the stringizing operator to save typing...
-#define	PSF(x) #x,(int)&((playerState_t*)0)->x
+#define	PSF(x) #x,(int)(intptr_t)&((playerState_t*)0)->x
 
 netField_t	playerStateFields[] = 
 {
